test: consolidate SQL executor Issue/GitHubIssue tests into thematic suites
Move @Test methods from 8 Issue/GitHubIssue source files in
engine/src/test/java/com/arcadedb/query/sql/executor/ into existing
thematic classes. Each migrated method carries a single-line
// Issue #NNNN comment for regression traceability.

Migrations:
- UpdateContentArrayTest (+3): GitHubIssue1315 (UPDATE/CREATE VERTEX
  with arrays)
- UpdateStatementExecutionTest (+10): 5 from GitHubIssue3813 (UPDATE
  subquery + RID, sqlscript LET update) + 5 from Issue1814
  (UPDATE CONTENT applyDefaults)
- DeleteExecutionStepTest (+1): GitHubIssue3813 deleteWhereRid
- ScriptExecutionTest (+2): GitHubIssue3871 (DELETE FROM variable in
  script)
- SelectStatementExecutionTest (+6): 4 from GitHubIssue3872 (LET map
  literal / subquery + trailing selector) + 2 from GitHubIssue3878
  (equality / IF in projection)
- MaterializedViewSQLTest (+2): GitHubIssue3878 (equality / IF in MV
  projection) - cross-package target under com.arcadedb.schema
- UnwindExpandStepTest (+7): 5 from Issue1582 (UNWIND projection) + 2
  from Issue2776 (EXPAND/UNWIND + LET variable)

Setup conflicts handled by inlining per-test type creation; moved
SQLScript update/delete tests slot into target classes with explicit
transaction wrappers matching existing style.

Verified: engine test-compile clean; 265 tests pass across the 7
modified target classes.
